requirements.txt
jq==0.1.6
PyYAML==5.1
requests==2.21.0
psycopg2==2.8.2
cleo==0.6.8
orator==0.9.8
Flask==1.0.2
flask_cors==3.0.7
flask_restplus==0.12.1
flask_orator==0.2.0
